05-full-projects/project-textual-project-planner/pyproject.toml
[project]
name = "flock-flightplan"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flock-core>=0.4.0b20",
    "textual>=3.1.0",
    "httpx>=0.26.0",
    "textual-dev>=1.7.0",
    "textual-web>=0.4.2",
    "radon>=6.0.1",
]

[project.scripts]
flock-flightplan = "flock_flightplan:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "debugpy>=1.8.14",
]
